backend/requirements.txt
fastapi
uvicorn
pymupdf
langchain
langchain-community
openai
chromadb
python-multipart
python-dotenv
